chore: move responsibility and resp_maps into new responsbility submodule
