+++ rex_xai/explanation.py
@@ -10,7 +10,7 @@ from rex_xai.input.config import CausalArgs, Strategy
 from rex_xai.input.input_data import Data
 from rex_xai.utils.logger import logger
 from rex_xai.mutants.mutant import _apply_to_data
-from rex_xai.resp_maps import ResponsibilityMaps
+from rex_xai.responsibility.resp_maps import ResponsibilityMaps
 
 
 class Explanation:
